Add files via upload
This Machine Learning Model is a Hybrid model which combines of the strengths of both Multi Layer Perceptron and Long Short Term Memory in detecting the source of leaf wetness using the parameters leaf moisture and leaf temperature. The steps involved in the development of the MLP-LSTM Hybrid Machine Learning model is present in these files.